DEV-120 refactor: rename `lib/highlights.lua` to `lib/hls.lua`
