+++ lua/terra-core/lib/bg.lua
@@ -1,5 +1,3 @@
-local conditional_hl = require("terra-core.lib.highlights").conditional_hl
-
 local M = {}
 
 ---@param config TerraConfig
@@ -9,7 +7,7 @@ function M.main(config, colors)
     local no_color = colors.none
     local bg = colors.semantics.bg
 
-    return conditional_hl(bg.primary.main, {
+    return require("terra-core.lib.hls").conditional_hl(bg.primary.main, {
         [config.transparency == "partial"] = no_color,
         [config.transparency == "full"] = no_color,
     })
@@ -24,7 +22,7 @@ function M.sidebar(config, colors)
     local bg = colors.semantics.bg
     local no_color = colors.none
 
-    return conditional_hl(bg.primary.dark, {
+    return require("terra-core.lib.hls").conditional_hl(bg.primary.dark, {
         [not dark_sidebars] = bg.primary.main,
         [dark_sidebars and transparency == "partial"] = bg.primary.dark,
         [transparency == "full"] = no_color,
@@ -38,7 +36,7 @@ end
 function M.float(config, colors)
     local bg = colors.semantics.bg
 
-    return conditional_hl(bg.primary.dark, {
+    return require("terra-core.lib.hls").conditional_hl(bg.primary.dark, {
         -- I decided to always use the dark background for floating windows
         -- But just in case I change my mind, I'll leave this here
     })
