+++ lua/terra-core/lib/init.lua
@@ -2,7 +2,7 @@ local lib = {
     config = require("terra-core.lib.config"),
     color = require("terra-core.lib.color"),
     ui = require("terra-core.lib.ui"),
-    hls = require("terra-core.lib.highlights"),
+    hls = require("terra-core.lib.hls"),
     files = require("terra-core.lib.files"),
     themes = require("terra-core.lib.themes"),
     debug = require("terra-core.lib.debug"),
